Reformatting and cleanup. Some variables renamed.

+++ g_spawn.c
@@ -537,7 +537,7 @@ char* ED_NewString(char* string)
 	char* newb, * new_p;
 	int		i, l;
 
-	l = strlen(string) + 1;
+	l = (int)strlen(string) + 1;
 
 	newb = gi.TagMalloc(l, TAG_LEVEL);
 
@@ -1200,7 +1200,7 @@ void SpawnEntities(char* mapname, char* entstring, char* spawnpoint)
 		char path[MAX_OSPATH];
 
 		// save original entstring for later
-		original_entstring = (char*)gi.TagMalloc(strlen(entstring), TAG_GAME);
+		original_entstring = (char*)gi.TagMalloc((int)strlen(entstring), TAG_GAME);
 		if (original_entstring == NULL)
 		{
 			gi.error("SpawnEntities: can't allocate original_entstring.\n");
